refactor(tests): consolidate FakeRedis implementations into tests/helpers/fake_redis.py (#5431)
