+++ autobot-backend/tests/test_knowledge_boards.py
@@ -15,7 +15,6 @@ The knowledge-base singleton is monkey-patched on each request via
 """
 
 import json
-from typing import Optional
 from unittest.mock import AsyncMock
 
 import pytest
@@ -28,42 +27,7 @@ from api.knowledge_boards import (
     GLOBAL_BOARD_ID,
 )
 from api.knowledge_boards import router as boards_router
-
-# ---------------------------------------------------------------------------
-# Fake in-process Redis hash
-# ---------------------------------------------------------------------------
-
-
-class _FakeRedis:
-    """Minimal async Redis stub that emulates hset / hget / hgetall / hdel."""
-
-    def __init__(self):
-        self._store: dict[str, dict[str, bytes]] = {}
-
-    async def hset(self, name: str, key: str, value: str) -> int:
-        bucket = self._store.setdefault(name, {})
-        existed = key in bucket
-        bucket[key] = value.encode("utf-8") if isinstance(value, str) else value
-        return 0 if existed else 1
-
-    async def hget(self, name: str, key: str) -> Optional[bytes]:
-        return self._store.get(name, {}).get(key)
-
-    async def hgetall(self, name: str) -> dict[bytes, bytes]:
-        raw = self._store.get(name, {})
-        return {
-            (k.encode("utf-8") if isinstance(k, str) else k): (
-                v if isinstance(v, bytes) else v.encode("utf-8")
-            )
-            for k, v in raw.items()
-        }
-
-    async def hdel(self, name: str, key: str) -> int:
-        bucket = self._store.get(name, {})
-        if key in bucket:
-            del bucket[key]
-            return 1
-        return 0
+from tests.helpers.fake_redis import AsyncHashFakeRedis
 
 
 # ---------------------------------------------------------------------------
@@ -71,7 +35,7 @@ class _FakeRedis:
 # ---------------------------------------------------------------------------
 
 
-def _make_app(fake_redis: _FakeRedis) -> FastAPI:
+def _make_app(fake_redis: AsyncHashFakeRedis) -> FastAPI:
     """Build a minimal FastAPI app with the boards router wired up."""
 
     class _FakeKB:
@@ -99,7 +63,7 @@ def _make_app(fake_redis: _FakeRedis) -> FastAPI:
 # ---------------------------------------------------------------------------
 
 
-def _seed_board(fake_redis: _FakeRedis, board_id: str, name: str) -> dict:
+def _seed_board(fake_redis: AsyncHashFakeRedis, board_id: str, name: str) -> dict:
     """Synchronously seed a board entry into the fake Redis store."""
     entry = {"board_id": board_id, "name": name, "description": "", "created_at": "2026-01-01T00:00:00+00:00"}
     # Direct store manipulation (sync path for test setup)
@@ -116,7 +80,7 @@ class TestKnowledgeBoardsList:
     """GET /boards always returns at least the global sentinel."""
 
     def test_empty_store_returns_global_board(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.get("/api/knowledge_base/boards")
         assert resp.status_code == 200
@@ -126,7 +90,7 @@ class TestKnowledgeBoardsList:
         assert GLOBAL_BOARD_ID in ids
 
     def test_seeded_boards_appear_in_list(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         _seed_board(fake_redis, "project-alpha", "Project Alpha")
         client = TestClient(_make_app(fake_redis))
         resp = client.get("/api/knowledge_base/boards")
@@ -138,7 +102,7 @@ class TestKnowledgeBoardsList:
         assert data["total"] == 2
 
     def test_total_count_includes_global(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         _seed_board(fake_redis, "board-a", "Board A")
         _seed_board(fake_redis, "board-b", "Board B")
         client = TestClient(_make_app(fake_redis))
@@ -150,7 +114,7 @@ class TestKnowledgeBoardsCreate:
     """POST /boards — creation, validation, and duplicate handling."""
 
     def test_create_board_returns_201(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.post(
             "/api/knowledge_base/boards",
@@ -162,7 +126,7 @@ class TestKnowledgeBoardsCreate:
         assert data["created"] is True
 
     def test_create_board_without_id_autogenerates(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.post(
             "/api/knowledge_base/boards",
@@ -174,7 +138,7 @@ class TestKnowledgeBoardsCreate:
         assert data["board_id"]  # non-empty
 
     def test_create_board_persists_in_redis(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         client.post(
             "/api/knowledge_base/boards",
@@ -186,7 +150,7 @@ class TestKnowledgeBoardsCreate:
         assert entry["name"] == "Persisted"
 
     def test_duplicate_board_id_returns_409(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         _seed_board(fake_redis, "existing-board", "Existing Board")
         client = TestClient(_make_app(fake_redis))
         resp = client.post(
@@ -197,7 +161,7 @@ class TestKnowledgeBoardsCreate:
 
     def test_reserved_global_board_id_returns_422(self):
         """__global__ is reserved; the validator should reject it."""
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.post(
             "/api/knowledge_base/boards",
@@ -208,7 +172,7 @@ class TestKnowledgeBoardsCreate:
 
     def test_invalid_board_id_chars_returns_422(self):
         """Board IDs must match [a-z0-9_-]."""
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.post(
             "/api/knowledge_base/boards",
@@ -217,7 +181,7 @@ class TestKnowledgeBoardsCreate:
         assert resp.status_code == 422
 
     def test_board_id_with_spaces_returns_422(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.post(
             "/api/knowledge_base/boards",
@@ -230,7 +194,7 @@ class TestKnowledgeBoardsDelete:
     """DELETE /boards/{board_id} — removal and guard cases."""
 
     def test_delete_existing_board_returns_200(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         _seed_board(fake_redis, "to-delete", "To Delete")
         client = TestClient(_make_app(fake_redis))
         resp = client.delete("/api/knowledge_base/boards/to-delete")
@@ -240,7 +204,7 @@ class TestKnowledgeBoardsDelete:
         assert data["board_id"] == "to-delete"
 
     def test_delete_removes_board_from_redis(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         _seed_board(fake_redis, "gone-board", "Gone Board")
         client = TestClient(_make_app(fake_redis))
         client.delete("/api/knowledge_base/boards/gone-board")
@@ -248,14 +212,14 @@ class TestKnowledgeBoardsDelete:
         assert remaining is None
 
     def test_delete_unknown_board_returns_404(self):
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.delete("/api/knowledge_base/boards/does-not-exist")
         assert resp.status_code == 404
 
     def test_delete_global_board_returns_400(self):
         """The __global__ sentinel must never be deletable."""
-        fake_redis = _FakeRedis()
+        fake_redis = AsyncHashFakeRedis()
         client = TestClient(_make_app(fake_redis))
         resp = client.delete(f"/api/knowledge_base/boards/{GLOBAL_BOARD_ID}")
         assert resp.status_code == 400
